+++ xapian-letor/ranklist.h
@@ -53,8 +53,8 @@ class XAPIAN_VISIBILITY_DEFAULT RankList {
     
     void set_qid(std::string qid1);
     
-    long int get_num_of_feature_vectors();
-    
+    int get_num_of_feature_vectors();
+
     void set_rl(std::vector<FeatureVector> local_rl);
 
     void add_feature_vector(const Xapian::FeatureVector fv);//was & fv initially,check back later
